third_party/WebKit/LayoutTests/external/wpt/html/dom/elements/the-innertext-idl-attribute/multiple-text-nodes.window-expected.txt

This is a testharness.js-based test.
FAIL Ensure multiple text nodes get rendered properly assert_equals: expected "X Y" but got "XY"
Harness: the test ran to completion.
